create user claim screen function
